cinegraph/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
neo4j==5.19.0
openai==1.30.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
streamlit==1.35.0
streamlit-agraph==0.0.45
